EmployeeDetails
Mycaption Project to "Write a C program that defines a structure employee containing the details such as empno, empname, department name and salary. The structure has to store 20 employees in an organization. Use the appropriate method to define the above details and define a function that will display the contents?"